% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the accuracy rate.}
\usage{
accuracy(observed, predicted)
}
\arguments{
\item{observed}{vector of observed values.}

\item{predicted}{vector of predicted values.}
}
\value{
the accuracy as a numeric value.
}
\description{
This function provides the overall accuracy rate for the two vectors.
}
