% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Marketing}
\format{
a data frame with 4,521 observations of 17 variables.
\describe{
    \item{age (numeric)}{age}
    \item{job}{type of job}
    \item{marital}{marital status (categorical: "married","divorced","single"; note: "divorced" means divorced or widowed)}
    \item{education}{educational level (categorical: "unknown","secondary","primary","tertiary")}
    \item{default}{has credit in default? (binary: "yes","no")}
    \item{balance}{average yearly balance, in euros (numeric)}
    \item{housing}{has housing loan? (binary: "yes","no")}
    \item{loan}{has personal loan? (binary: "yes","no")}
    \item{contact}{contact communication type (categorical: "unknown","telephone","cellular")}
    \item{day}{last contact day of the month (numeric)}
    \item{month}{last contact month of year (categorical: "jan", "feb", "mar", ..., "nov", "dec")}
    \item{duration}{last contact duration, in seconds (numeric)}
    \item{campaign}{number of contacts performed during this campaign and for this client (numeric, includes last contact)}
    \item{pdays}{number of days that passed by after the client was last contacted from a previous campaign (numeric, -1 means client was not previously contacted)}
    \item{previous}{number of contacts performed before this campaign and for this client (numeric)}
    \item{poutcome}{outcome of the previous marketing campaign (categorical: "unknown","other","failure","success")}
    \item{subscribed}{has the client subscribed a term deposit? (binary: "yes","no")}
}
}
\source{
https://archive.ics.uci.edu/dataset/222/bank+marketing
}
\description{
The data is related with direct marketing campaigns (phone calls) of a Portuguese banking
institution. The classification goal is to predict if the client will subscribe a term deposit
(variable `subscribed`).
}
