% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{calculate_roc}
\alias{calculate_roc}
\alias{summary.roc}
\alias{plot.roc}
\title{Calculate the statistics for receiver operating characteristic curve}
\usage{
calculate_roc(predictions, observed, cost_of_fp = 1, cost_of_fn = 1, n = 100)

\method{summary}{roc}(object, digits = 3, ...)

\method{plot}{roc}(x, curve = "accuracy", legend.position = c(1, 0.2), ...)
}
\arguments{
\item{predictions}{predicted values.}

\item{observed}{actual observed outcomes.}

\item{cost_of_fp}{cost of a false positive.}

\item{cost_of_fn}{cost of a false negative.}

\item{n}{the number of steps to estimate.}

\item{object}{result of [calculate_roc()].}

\item{digits}{number of digits to print.}

\item{...}{currently unused.}

\item{x}{result of [calculate_roc()].}

\item{curve}{values can be cost, accuracy, or NULL.}

\item{legend.position}{position of the legend for teh accuracy curve plot.}
}
\value{
a ggplot2 expression.
}
\description{
This function was adapted from Raffel (https://github.com/joyofdata):
https://github.com/joyofdata/joyofdata-articles/blob/master/roc-auc/calculate_roc.R
}
