\name{combine.medley}
\alias{combine.medley}
\title{Recursively combine two or more medley objects for the same problem}
\usage{
  combine.medley(e1, e2 = NULL, ...)
}
\arguments{
  \item{e1}{the first medley to combine}

  \item{e2}{the second medley to combine}

  \item{...}{further medleys to combine}
}
\description{
  Recursively combine two or more medley objects for the
  same problem
}

