% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{combine_confusion_matrices}
\alias{combine_confusion_matrices}
\title{Combine multiple confusion matrices into a single table.}
\usage{
combine_confusion_matrices(..., digits = 2)
}
\arguments{
\item{...}{results from [confusion_matrix()]. If the parameters are named
those will be used as the row names for each matrix.}

\item{digits}{number of digits.}
}
\description{
This will combine multiple confusion matrices into a single table ideal
for comparing the performance across multiple models.
}
