\name{constant}
\alias{constant}
\title{Trivial regression model which returns a given constant}
\usage{
  constant(x, y, k)
}
\arguments{
  \item{x}{the matrix of predictors}

  \item{y}{the vector of responses}

  \item{k}{the constant to return}
}
\description{
  Trivial regression model which returns a given constant
}

