% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Contingency and proportional table as a character}
\usage{
crosstab(x, digits = 2, useNA = "ifany", sep = "; ")
}
\arguments{
\item{x}{a factor.}

\item{digits}{number of digits for proportions.}

\item{useNA}{whether to include NA values in the table. See details in [base::table()].}

\item{sep}{separator between categories.}
}
\value{
a character.
}
\description{
This is a wrapper to [base::table()] and [base::prop.table()] but returns the contingency and
proportional table as a single character string.
}
\examples{
data(mtcars)
crosstab(mtcars$am)
}
