% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daacs_data.R
\docType{data}
\name{daacs}
\alias{daacs}
\title{Data from the Diagnostic Assessment and Achievement of College Skills (DAACS)}
\format{
## `daacs`
A data frame with 5,154 observations of 14 variables.
\describe{
    \item{retained}{whether the student was retained for a second semester}
    \item{srl}{the student's self-regulated learning score}
    \item{math}{the student's mathematics score}
    \item{reading}{the student's reading score}
    \item{writing}{the student's writing score}
    \item{income}{the student's income level as an ordered factor}
    \item{employment}{the student's employment status}
    \item{ell}{whether the student is an English Language Learner}
    \item{ed_mother}{the highest education level of the student's mother}
    \item{ed_father}{the highest education level of the student's father}
    \item{ethnicity}{the ethnicity of the student}
    \item{gender}{the gender of the student}
    \item{military}{whether the student is active in the military}
    \item{age}{the age of the student at the time of enrollment}
}
}
\source{
https://daacs.net
}
\usage{
daacs
}
\description{
This is data part of a larger randomized control trial designed to test
the effect of DAACS on student success. These data are from the treatment
group at a large online college. DAACS was embedded within orientation.
Although students were expected to complete orientation there were no
consequences for not doing so. As a result some or all DAACS data are missing
for a large proportion of the students.
}
\details{
For more information about DAACS, see https://daacs.net.
}
\keyword{datasets}
