% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_formula.R
\name{expand_formula}
\alias{expand_formula}
\title{Utility function that will expand the formula if the independent variable(s)
are not full specified.}
\usage{
expand_formula(formula, data)
}
\arguments{
\item{formula}{the formula to expand.}

\item{data}{the data.frame the formula will be applied to.}
}
\value{
a formula with all independent variables specified explicitly.
}
\description{
This function will add all the independent variables to the formula if the
formula was specified as `y ~ .`.
}
