% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varaible_sets.R
\name{get_variable_sets}
\alias{get_variable_sets}
\alias{print.variableset}
\title{Gets list of model formulas based upon the missing data pattern in the data set.}
\usage{
get_variable_sets(data, formula, min_set_size = 0.1)

\method{print}{variableset}(x, ...)
}
\arguments{
\item{data}{the data.frame.}

\item{formula}{the formula that includes all the possible variables to consider
for the predictive models.}

\item{min_set_size}{the minimum set size as a percentage to include as a model.}

\item{x}{the results from `get_variable_sets()`.}

\item{...}{currently unused.}
}
\value{
a list of formulas, in order to be used, for the predictive models.
}
\description{
Gets list of model formulas based upon the missing data pattern in the data set.
}
