\name{logloss}
\alias{logloss}
\title{Calculate log-loss error metric}
\usage{
  logloss(true, pred, eps = 1e-15)
}
\arguments{
  \item{true}{the true response values}

  \item{pred}{the predicted response values}

  \item{eps}{a lower bound on predicted probabilities}
}
\description{
  Calculate log-loss error metric
}

