% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{accuracy}
\alias{accuracy}
\alias{sensitivity}
\alias{specificity}
\title{Calculate the accuracy rate.}
\usage{
accuracy(observed, predicted)

sensitivity(observed, predicted)

specificity(observed, predicted)
}
\arguments{
\item{observed}{vector of observed values.}

\item{predicted}{vector of predicted values.}
}
\value{
the accuracy as a numeric value.
}
\description{
This function provides the overall accuracy rate for the two vectors. This is the sum of the
true positive and true negative values.

sum(True positive) / sum(Condition positive)

sum(True negative) / sum(Condition negative)
}
