% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhis_data.R
\docType{data}
\name{nhis}
\alias{nhis}
\title{National Health Interview Survey (NHIS) 2017}
\format{
a data frame with 26,742 observations of 1,414 variables.
}
\source{
https://archive.cdc.gov/www_cdc_gov/nchs/nhis/nhis_2017_data_release.htm
}
\description{
The National Health Interview Survey (NHIS) monitors the health of the civilian
non-institutionalized U.S. population through the collection and analysis of data on a broad
range of health topics.
}
\details{
This dataset was used by Perez-Lebel (2022) in their paper Benchmarking missing-values
approaches for predictive models on health databases." More information is available here:
https://github.com/aperezlebel/benchmark_mv_approaches
}
