% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa}
\alias{pisa}
\title{Programme of International Student Assessment}
\format{
A data frame with 5,233 observations of 45 variables.
\describe{
    \item{Public}{Whether student attends a public or private school.}
    \item{ST04Q01}{Sex}
    \item{ST05Q01}{Attend <ISCED 0>}
    \item{ST06Q01}{Age at <ISCED 1>}
    \item{ST07Q01}{Repeat <ISCED 1>}
    \item{ST08Q01}{At Home - Mother}
    \item{ST08Q02}{At Home - Father}
    \item{ST08Q03}{At Home - Brothers}
    \item{ST08Q04}{At Home - Sisters}
    \item{ST08Q05}{At Home - Grandparents}
    \item{ST08Q06}{At Home - Others}
    \item{ST10Q01}{Mother  <Highest Schooling>}
    \item{ST12Q01}{Mother Current Job Status}
    \item{ST14Q01}{Father  <Highest Schooling>}
    \item{ST16Q01}{Father Current Job Status}
    \item{ST19Q01}{Language at home}
    \item{ST20Q01}{Possessions desk}
    \item{ST20Q02}{Possessions own room}
    \item{ST20Q03}{Possessions study place}
    \item{ST20Q04}{Possessions  computer}
    \item{ST20Q05}{Possessions software}
    \item{ST20Q06}{Possessions Internet}
    \item{ST20Q07}{Possessions literature}
    \item{ST20Q08}{Possessions poetry}
    \item{ST20Q09}{Possessions art}
    \item{ST20Q10}{Possessions textbooks}
    \item{ST20Q12}{Possessions dictionary}
    \item{ST20Q13}{Possessions dishwasher}
    \item{ST21Q01}{How many cellular phones}
    \item{ST21Q02}{How many televisions}
    \item{ST21Q03}{How many computers}
    \item{ST21Q04}{How many cars}
    \item{ST21Q05}{How many rooms bath or shower}
    \item{ST22Q01}{How many books at home}
    \item{ST23Q01}{Reading Enjoyment Time}
    \item{ST31Q01}{<Enrich> in <test lang>}
    \item{ST31Q02}{<Enrich> in <mathematics>}
    \item{ST31Q03}{<Enrich> in <science>}
    \item{ST31Q05}{<Remedial> in <test lang>}
    \item{ST31Q06}{<Remedial> in <mathematics>}
    \item{ST31Q07}{<Remedial> in <science>}
    \item{ST32Q01}{Out of school lessons <test lang>}
    \item{ST32Q02}{Out of school lessons <maths>}
    \item{ST32Q03}{Out of school lessons <science>}
}
}
\source{
https://www.pisa.oecd.org
}
\description{
This data is from the 2019 PISA implementation for the United States. It is used to predict
whether a student attends a public or private school. The dependent variable (`Public`) is
is imbalanced with approximately 93% of students attending public schools.
}
\details{
This dataset was modified from the original data provided by OECD.
The [`pisa`](https://github.com/jbryer/pisa) R package on Github provides the complete data
for the 2019 administration.
}
