% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa_variables}
\alias{pisa_variables}
\title{Programme of International Student Assessment}
\format{
a character vector.
}
\source{
https://www.pisa.oecd.org
}
\description{
This is a character vector where the names of the variables correspond to the variables names
in the `pisa` data frame and the values are descriptions of the variables.
}
