\name{predict.constant}
\alias{predict.constant}
\title{Trivial regression model which returns a given constant}
\usage{
  \method{predict}{constant} (object, newdata, ...)
}
\arguments{
  \item{object}{the model object}

  \item{newdata}{the matrix of new predictors}

  \item{...}{other parameters (unused)}
}
\description{
  Trivial regression model which returns a given constant
}

