% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_small_levels.R
\name{recode_small_levels}
\alias{recode_small_levels}
\title{Recode levels of a factor that are below a given threshold.}
\usage{
recode_small_levels(x, recode_to = NA, threshold = length(x) * 0.01, ...)
}
\arguments{
\item{x}{a vector.}

\item{recode_to}{the value to recode small factor levels to.}

\item{threshold}{the threshold of a factor level count for it to be recoded.}

\item{...}{other parameters passed to `factor()`.}
}
\value{
a recoded factor. The result will be a factor regardless of the type of `x`.
}
\description{
This function will first calculate a contingency table. For any categories that are below
the specified `threshold`, the raw values in `x` will be recoded. The default is `NA`, but can
be specified by the `recode_to` parameter. The resulting vector will be converted to a factor
using the `factor()` function.
}
\examples{
data(mtacars)
table(mtcars$carb)
cbind(mtcars$carb, medley::recode_small_levels(mtcars$carb, threshold = 2))
}
