% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\alias{as.data.frame.confusionmatrix}
\alias{print.confusionmatrix}
\title{Calculate a confusion matrix}
\usage{
confusion_matrix(observed, predicted)

\method{as.data.frame}{confusionmatrix}(x, row.names = NULL, optional = FALSE, digits = 2, ...)

\method{print}{confusionmatrix}(x, digits = 2, ...)
}
\arguments{
\item{observed}{vector of observed values.}

\item{predicted}{vector of predicted values.}

\item{x}{the result of [confusion_matrix()].}

\item{row.names}{not used.}

\item{optional}{not used.}

\item{digits}{number of decimal places to print.}

\item{...}{currently not used.}
}
\value{
a data.frame of the confusion matrix.
}
\description{
Calculate a confusion matrix
}
\examples{
observed <- c(rep(FALSE, 10), rep(TRUE, 2))
predicted <- rep(TRUE, 12)
confusion_matrix(observed, predicted)
}
