% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medley.R
\name{medley}
\alias{medley}
\alias{summary.medley}
\alias{print.medley}
\alias{predict.medley}
\title{Train models using different combinations of predictor variables based upon
missing data patterns.}
\usage{
medley(
  data,
  formula,
  method = glm,
  var_sets = get_variable_sets(data = data, formula = formula, min_set_size =
    min_set_size),
  min_set_size = 0.1,
  exclusive_membership = TRUE,
  ...
)

\method{summary}{medley}(object, ...)

\method{print}{medley}(x, ...)

\method{predict}{medley}(object, newdata, ...)
}
\arguments{
\item{data}{data.frame used to estimate the models.}

\item{formula}{with all possible predictor varaibles to be considered.}

\item{method}{the function used to train the models (e.g. glm, randomForest).}

\item{var_sets}{a list of formulas to use for the predictive models.}

\item{min_set_size}{the minimum set size as a percentage to incldue as a model.}

\item{exclusive_membership}{whether an observation should only be used only in
the model for which the most predictor variables are available. If
`FALSE` then observations may be used in training more than one model.}

\item{...}{other parameters passed to the `predict()` function.}

\item{object}{the results from `medley`.}

\item{x}{the results of `medley`.}

\item{newdata}{(optional) a new data.frame to get predictions for.}
}
\value{
an object with the following elements:
\describe{
\item{n_models}{the number of models trained.}
\item{formulas}{the list of formulas used to train the models.}
\item{models}{list of objects returned from the training method.}
\item{data}{the data.frame used to train the models.}
\item{model_observations}{a data.frame that specifies which observations are used for which model(s).}
}

a vector of predictions.
}
\description{
Train models using different combinations of predictor variables based upon
missing data patterns.
}
